+++ ExcelDna.CellAddress.sln
@@ -4,8 +4,6 @@ VisualStudioVersion = 15.0.27130.2036
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExcelDna.CellAddress", "src\ExcelDna.CellAddress\ExcelDna.CellAddress.csproj", "{83F84764-54EB-4B6C-8E1E-302EC42F83D2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CellAddress.Tests.AddIn", "test\CellAddress.Tests.AddIn\CellAddress.Tests.AddIn.csproj", "{E4CBBE63-5DD2-4B2A-A1AE-5CCBFA5B055C}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CellAddress.Tests", "test\CellAddress.Tests\CellAddress.Tests.csproj", "{8836D7C7-0439-41C4-9D70-127C3CB27A42}"
 EndProject
 Global
@@ -18,10 +16,6 @@ Global
 		{83F84764-54EB-4B6C-8E1E-302EC42F83D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{83F84764-54EB-4B6C-8E1E-302EC42F83D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{83F84764-54EB-4B6C-8E1E-302EC42F83D2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E4CBBE63-5DD2-4B2A-A1AE-5CCBFA5B055C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E4CBBE63-5DD2-4B2A-A1AE-5CCBFA5B055C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E4CBBE63-5DD2-4B2A-A1AE-5CCBFA5B055C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E4CBBE63-5DD2-4B2A-A1AE-5CCBFA5B055C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8836D7C7-0439-41C4-9D70-127C3CB27A42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8836D7C7-0439-41C4-9D70-127C3CB27A42}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8836D7C7-0439-41C4-9D70-127C3CB27A42}.Release|Any CPU.ActiveCfg = Release|Any CPU
