+++ src/ExcelDna.CellAddress/ExcelDna.CellAddress.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net40</TargetFramework>
+    <TargetFrameworks>net40;net451</TargetFrameworks>
     <AssemblyName>ExcelDna.CellAddress</AssemblyName>
     <RootNamespace>ExcelDna</RootNamespace>
     <VersionPrefix>1.0.1</VersionPrefix>
